requirements.txt
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
streamlit>=1.33.0
python-dotenv>=1.0.0
PyGithub>=2.1.0
sentence-transformers>=2.2.0
pydantic-settings>=2.0.0
pytest>=7.0.0
numpy>=1.24.0
streamlit-autorefresh>=0.0.1
